test: Move project and folder tests to projects/ directory
