+++ packages/testing/playwright/tests/ui/projects/folders-advanced.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../fixtures/base';
 
 test.describe('Folders - Advanced Operations', () => {
 	test.describe('Duplicate workflows', () => {
